create main menu
